+++ src/Handler/PcntlForkingHandler.php
@@ -13,9 +13,13 @@
 
 namespace PMG\Queue\Handler;
 
+use GuzzleHttp\Promise\Promise;
+use GuzzleHttp\Promise\PromiseInterface;
 use PMG\Queue\Message;
 use PMG\Queue\MessageHandler;
 use PMG\Queue\Exception\CouldNotFork;
+use PMG\Queue\Exception\ForkedProcessCancelled;
+use PMG\Queue\Exception\ForkedProcessFailed;
 
 /**
  * A message handler decorator that forks a child process to handle the message.
@@ -54,18 +58,38 @@ final class PcntlForkingHandler implements MessageHandler
      * `MessageHandler`. Just be sure to return `false` (the job failed) so it
      * can be retried.
      */
-    public function handle(Message $message, array $options=[])
+    public function handle(Message $message, array $options=[]) : PromiseInterface
     {
+        // this is outside the promise so both the cancel and wait
+        // callbacks have access to the child's PID
         $child = $this->fork();
         if (0 === $child) {
             try {
-                $result = $this->wrapped->handle($message, $options);
+                $result = $this->wrapped->handle($message, $options)->wait();
             } finally {
                 $this->pcntl->quit(isset($result) && $result);
             }
         }
 
-        return $this->pcntl->wait($child);
+        $promise = new Promise(function () use (&$promise, $child) {
+            $succeeded = $this->pcntl->wait($child);
+            // this happens when the promise is cancelled. We don't want to
+            // to try and change the promise to resolved if that happens.
+            if ($promise->getState() !== PromiseInterface::PENDING) {
+                return;
+            }
+
+            if ($succeeded) {
+                $promise->resolve(true);
+            } else {
+                $promise->reject(new ForkedProcessFailed());
+            }
+        }, function () use (&$promise, $child) {
+            $this->pcntl->signal($child, SIGTERM);
+            $promise->reject(new ForkedProcessCancelled());
+        });
+
+        return $promise;
     }
 
     private function fork()