+++ test/unit/DefaultConsumerTest.php
@@ -13,6 +13,7 @@
 
 namespace PMG\Queue;
 
+use GuzzleHttp\Promise\FulfilledPromise;
 use Psr\Log\LogLevel;
 use PMG\Queue\Exception\SimpleMustStop;
 
@@ -43,7 +44,7 @@ class DefaultConsumerTest extends UnitTestCase
         $this->handler->expects($this->once())
             ->method('handle')
             ->with($this->identicalTo($this->message))
-            ->willReturn(true);
+            ->willReturn(self::promise(true));
 
         $this->assertTrue($this->consumer->once(self::Q));
     }
@@ -58,7 +59,7 @@ class DefaultConsumerTest extends UnitTestCase
         $this->handler->expects($this->once())
             ->method('handle')
             ->with($this->identicalTo($this->message))
-            ->willReturn(false);
+            ->willReturn(self::promise(false));
 
         $this->assertFalse($this->consumer->once(self::Q));
     }
@@ -74,7 +75,7 @@ class DefaultConsumerTest extends UnitTestCase
         $this->handler->expects($this->once())
             ->method('handle')
             ->with($this->identicalTo($this->message))
-            ->willReturn(false);
+            ->willReturn(self::promise(false));
 
         $this->assertFalse($this->consumer->once(self::Q));
     }
@@ -114,6 +115,20 @@ class DefaultConsumerTest extends UnitTestCase
         $this->consumer->once(self::Q);
     }
 
+    public function testFailureWithShouldReleaseReleasesMessageBackIntoDriver()
+    {
+        $this->withMessage();
+        $this->driver->expects($this->once())
+            ->method('release')
+            ->with(self::Q, $this->envelope);
+        $this->handler->expects($this->once())
+            ->method('handle')
+            ->with($this->identicalTo($this->message))
+            ->willThrowException(new Exception\ForkedProcessCancelled('oops'));
+
+        $result = $this->consumer->once(self::Q);
+    }
+
     /**
      * @group https://github.com/AgencyPMG/Queue/issues/61
      */
@@ -133,7 +148,7 @@ class DefaultConsumerTest extends UnitTestCase
         $this->handler->expects($this->once())
             ->method('handle')
             ->with($this->identicalTo($this->message))
-            ->willReturn(true);
+            ->willReturn(self::promise(true));
 
         $result = $this->consumer->once(self::Q, $lifecycle);
 
@@ -160,7 +175,7 @@ class DefaultConsumerTest extends UnitTestCase
         $this->handler->expects($this->once())
             ->method('handle')
             ->with($this->identicalTo($this->message))
-            ->willReturn(false);
+            ->willReturn(self::promise(false));
 
         $result = $this->consumer->once(self::Q, $lifecycle);
 
@@ -191,5 +206,10 @@ class DefaultConsumerTest extends UnitTestCase
         $this->retries->expects($this->atLeastOnce())
             ->method('canRetry')
             ->willReturn(true);
+    }
+
+    private static function promise(bool $result) : FulfilledPromise
+    {
+        return new FulfilledPromise($result);
     }
 }