+++ test/unit/Handler/PcntlForkingHandlerTest.php
@@ -15,6 +15,8 @@ namespace PMG\Queue\Handler;
 
 use PMG\Queue\SimpleMessage;
 use PMG\Queue\Exception\CouldNotFork;
+use PMG\Queue\Exception\ForkedProcessCancelled;
+use PMG\Queue\Exception\ForkedProcessFailed;
 
 /**
  * This uses `CallableHandler` simply because I'm not sure how phpunit mock objects
@@ -34,45 +36,51 @@ class PcntlForkingHandlerTest extends \PMG\Queue\UnitTestCase
             return true;
         });
 
-        $this->assertTrue($handler->handle($this->message));
+        $promise = $handler->handle($this->message);
+
+        $this->assertTrue($promise->wait());
     }
 
-    public function testChildProcessWithFailedResultReturnsFalseFromTheParent()
+    public function testChildProcessWithFailedResultCausesErrorInParent()
     {
+        $this->expectException(ForkedProcessFailed::class);
         $handler = $this->createHandler(function () {
             return false;
         });
 
-        $this->assertFalse($handler->handle($this->message));
+        $handler->handle($this->message)->wait();
     }
 
     public function testChildProcessThatExitsEarlyWithErrorReturnsFalseFromParent()
     {
+        $this->expectException(ForkedProcessFailed::class);
         $handler = $this->createHandler(function () {
             // we can't throw here because php unit complains. Instead just call
             // `exit` with 255 to simulate exiting early.
             exit(255);
         });
 
-        $this->assertFalse($handler->handle($this->message));
+        $handler->handle($this->message)->wait();
     }
 
     public function testChildProcessThatThrowsAnExceptionExitsUnsuccessfully()
     {
+        $this->expectException(ForkedProcessFailed::class);
         $handler = $this->createHandler(function () {
             throw new \Exception('oh noz');
         });
 
-        $this->assertFalse($handler->handle($this->message));
+        $handler->handle($this->message)->wait();
     }
 
     public function testChildProcessWithErrorExitsUnsuccessfully()
     {
+        $this->expectException(ForkedProcessFailed::class);
         $handler = $this->createHandler(function () {
             throw new \Error('oh noz');
         });
 
-        $this->assertFalse($handler->handle($this->message));
+        $handler->handle($this->message)->wait();
     }
 
     public function testChildProcessIsPassedTheOptionsFromTheHandler()
@@ -83,7 +91,9 @@ class PcntlForkingHandlerTest extends \PMG\Queue\UnitTestCase
             return true;
         });
 
-        $this->assertTrue($handler->handle($this->message, ['one' => true]));
+        $promise = $handler->handle($this->message, ['one' => true]);
+
+        $this->assertTrue($promise->wait());
     }
 
     public function testHandlerErrorsIfAChildProcessCannotFork()
@@ -98,6 +108,46 @@ class PcntlForkingHandlerTest extends \PMG\Queue\UnitTestCase
             ->willReturn(-1);
 
         $handler->handle($this->message);
+    }
+
+    public function testHandlerPromisesCanBeCancelled()
+    {
+        $this->expectException(ForkedProcessCancelled::class);
+        $handler = $this->createHandler(function () {
+            sleep(10000);
+            $this->assertTrue(false, 'causes a different exit exception');
+        });
+
+        $promise = $handler->handle($this->message);
+        $promise->cancel();
+
+        $promise->wait();
+    }
+
+    /**
+     * @requires function pcntl_async_signals
+     * @group slow
+     */
+    public function testHandlersWaitingThatAreCancelledAsynchronouslyFail()
+    {
+        $this->expectException(ForkedProcessCancelled::class);
+        $handler = $this->createHandler(function () {
+            sleep(10000);
+            $this->assertTrue(false, 'causes a different exit exception');
+        });
+        $promise = $handler->handle($this->message);
+
+        $previous = pcntl_async_signals(true);
+        pcntl_signal(SIGALRM, function () use ($promise) {
+            $promise->cancel();
+        });
+        pcntl_alarm(3);
+        try {
+            $promise->wait();
+        } finally {
+            pcntl_async_signals($previous);
+            pcntl_signal(SIGALRM, SIG_DFL);
+        }
     }
 
     protected function setUp()